requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==3.0.1
asyncpg==0.25.0
attrs==21.2.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
discord==1.7.3
discord.py==1.7.3
Django==4.0
frozenlist==1.2.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-pubsub==2.9.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
html2text==2020.1.16
httplib2==0.20.2
idna==3.3
libcst==0.4.1
lxml==4.7.1
multidict==5.2.0
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.1.1
proto-plus==1.20.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
typing-inspect==0.7.1
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
yarl==1.7.2
